Removed derived fields from Diver entity:
- totalDives no longer persisted
- rank derived from computed dive count
- count retrieved via DiveLogRepository

+++ src/main/java/com/lucap/scubakeep/dto/DiverResponseDTO.java
@@ -34,7 +34,7 @@ public class DiverResponseDTO {
 
     private Role role;
 
-    private int totalDives;
+    private long totalDives;
     private Certification highestCertification;
     private Set<String> specialties;
     private String rank;
